requirements.txt
streamlit
groq
fpdf2
sqlparse
psycopg2-binary